Taxes_Managements/Backend/requirements.txt
bcrypt==4.0.1
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pymysql
cryptography
